Move test cases and apply code suggestion
