+++ src/Stryker.Core/Stryker.Core.UnitTest/Mutators/IsPatternExpressionMutatorTests.cs
@@ -41,56 +41,6 @@ public class IsPatternExpressionMutatorTests : TestBase
     }
 
     [TestMethod]
-    [DataRow(">", new[] { SyntaxKind.LessThanToken, SyntaxKind.GreaterThanEqualsToken })]
-    [DataRow("<", new[] { SyntaxKind.GreaterThanToken, SyntaxKind.LessThanEqualsToken })]
-    [DataRow(">=", new[] { SyntaxKind.GreaterThanToken, SyntaxKind.LessThanToken })]
-    [DataRow("<=", new[] { SyntaxKind.GreaterThanToken, SyntaxKind.LessThanToken })]
-    public void ShouldMutateRelationalPattern(string @operator, SyntaxKind[] mutated)
-    {
-        var target = new RelationalPatternMutator();
-
-        var expression = GenerateWithRelationalPattern(@operator).DescendantNodes().OfType<RelationalPatternSyntax>().First();
-
-        var result = target.ApplyMutations(expression, null).ToList();
-
-        result.ForEach(mutation =>
-        {
-            mutation.OriginalNode.ShouldBeOfType<RelationalPatternSyntax>();
-            mutation.ReplacementNode.ShouldBeOfType<RelationalPatternSyntax>();
-            mutation.DisplayName.ShouldBe($"Equality mutation");
-        });
-
-        result
-            .Select(mutation => (RelationalPatternSyntax)mutation.ReplacementNode)
-            .Select(pattern => pattern.OperatorToken.Kind())
-            .ShouldBe(mutated, true);
-    }
-
-    [TestMethod]
-    [DataRow("and", new[] { SyntaxKind.OrPattern })]
-    [DataRow("or", new[] { SyntaxKind.AndPattern })]
-    public void ShouldMutateLogicalPattern(string @operator, SyntaxKind[] mutated)
-    {
-        var target = new BinaryPatternMutator();
-
-        var expression = GenerateWithBinaryPattern(@operator);
-
-        var result = target.ApplyMutations(expression, null).ToList();
-
-        result.ForEach(mutation =>
-        {
-            mutation.OriginalNode.ShouldBeOfType<BinaryPatternSyntax>();
-            mutation.ReplacementNode.ShouldBeOfType<BinaryPatternSyntax>();
-            mutation.DisplayName.ShouldBe($"Logical mutation");
-        });
-
-        result
-            .Select(mutation => (BinaryPatternSyntax)mutation.ReplacementNode)
-            .Select(pattern => pattern.Kind())
-            .ShouldBe(mutated, true);
-    }
-
-    [TestMethod]
     [DynamicData(nameof(GenerateNotSupportedPatterns))]
     public void ShouldNotMutateNotSupportedPatterns(IsPatternExpressionSyntax expression)
     {
@@ -119,52 +69,6 @@ namespace TestApplication
         var isPatternExpression = tree.GetRoot()
             .DescendantNodes()
             .OfType<IsPatternExpressionSyntax>()
-            .Single();
-
-        return isPatternExpression;
-    }
-
-    private IsPatternExpressionSyntax GenerateWithRelationalPattern(string @operator)
-    {
-        var tree = CSharpSyntaxTree.ParseText($@"
-using System;
-
-namespace TestApplication
-{{
-    class Program
-    {{
-        static void Main(string[] args)
-        {{
-            var a = 1 is ({@operator} 1);
-        }}
-    }}
-}}");
-        var isPatternExpression = tree.GetRoot()
-            .DescendantNodes()
-            .OfType<IsPatternExpressionSyntax>()
-            .Single();
-
-        return isPatternExpression;
-    }
-
-    private BinaryPatternSyntax GenerateWithBinaryPattern(string pattern)
-    {
-        var tree = CSharpSyntaxTree.ParseText($@"
-using System;
-
-namespace TestApplication
-{{
-    class Program
-    {{
-        static void Main(string[] args)
-        {{
-            var a = 1 is (1 {pattern} 2);
-        }}
-    }}
-}}");
-        var isPatternExpression = tree.GetRoot()
-            .DescendantNodes()
-            .OfType<BinaryPatternSyntax>()
             .Single();
 
         return isPatternExpression;
